corrigido, 19 testes com documentacao doxygen

+++ readSVG.cpp
@@ -102,17 +102,17 @@ namespace svg
                 svg_elements.push_back(new Polygon(parse_color(fill_color), points));
             }
 
-            else if (strcmp(child->Name(), "rect") == 0)
+           else if (strcmp(child->Name(), "rect") == 0)
             {
-                float x = child->FloatAttribute("x");
-                float y = child->FloatAttribute("y");
-                float width = child->FloatAttribute("width");
-                float height = child->FloatAttribute("height");
+                // Read rectangle attributes
+                int x = child->FloatAttribute("x");
+                int y = child->FloatAttribute("y");
+                int width = child->FloatAttribute("width");
+                int height = child->FloatAttribute("height");
                 const char *fill_color = child->Attribute("fill");
-                
-                svg_elements.push_back(new Rect(parse_color(fill_color), {static_cast<int>(x), static_cast<int>(y)}, static_cast<int>(width), static_cast<int>(height)));
+                // Create Rectangle object and add to vector
+                svg_elements.push_back(new Rect(parse_color(fill_color), {x,y}, width, height));
             }
-
             
             child = child->NextSiblingElement();
         }
